fix usecases
