+++ lib/app/features/community/presentation/widgets/grouping_explanation.dart
