--- lib/app/features/groups/data/database/groups_remote_database.dart
@@ -1,115 +1,0 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:organizer_client/app/features/groups/domain/entities/group_entity.dart';
-import 'package:organizer_client/app/features/groups/domain/entities/group_member_entity.dart';
-import 'package:organizer_client/app/features/groups/domain/entities/sub_group_entity.dart';
-
-abstract class GroupsRemoteDatabase {
-  Future<String> createGroup(GroupEntity group);
-  Future<GroupEntity> findGroup(String groupId);
-  Future<SubGroupEntity> findSubGroup(String subGroupId);
-  Future<void> joinGroup({
-    required String subGroupId,
-    required GroupMemberEntity member,
-  });
-  Future<List<GroupEntity>> findCreatedGroups(String userId);
-  Future<List<SubGroupEntity>> findJoinedGroups(GroupMemberEntity member);
-}
-
-class GroupsRemoteDatabaseImpl implements GroupsRemoteDatabase {
-  @override
-  Future<String> createGroup(GroupEntity group) async {
-    // remove subgroups from group
-    final groupWithoutSubGroups = group.copyWith(subGroups: []);
-    // remove subgroup from group property
-
-    await FirebaseFirestore.instance
-        .collection('groups')
-        .doc(group.id)
-        .set(groupWithoutSubGroups.toMap());
-    // add subgroups to subgroups collection
-    for (var subGroup in group.subGroups) {
-      await FirebaseFirestore.instance
-          .collection('subGroups')
-          .doc(subGroup.id)
-          .set(subGroup.toMap());
-    }
-
-    return group.id;
-  }
-
-  @override
-  Future<void> joinGroup({
-    required String subGroupId,
-    required GroupMemberEntity member,
-  }) async {
-    await FirebaseFirestore.instance
-        .collection('subGroups')
-        .doc(subGroupId)
-        .update({
-      "members": FieldValue.arrayUnion([member.toMap()])
-    });
-  }
-
-  @override
-  Future<GroupEntity> findGroup(String groupId) async {
-    final group = await FirebaseFirestore.instance
-        .collection('groups')
-        .doc(groupId)
-        .get();
-    if (group.exists) {
-      final subGroups = await FirebaseFirestore.instance
-          .collection('subGroups')
-          .where('groupRef', isEqualTo: groupId)
-          .get();
-      final subGroupsList =
-          subGroups.docs.map((e) => SubGroupEntity.fromMap(e.data())).toList();
-      final groupData = group.data()!;
-      final List<GroupMemberEntity> allMembers = [];
-      for (var element in subGroupsList) {
-        allMembers.addAll(element.members);
-      }
-      groupData['subGroups'] = subGroupsList;
-      groupData['members'] = allMembers;
-      return GroupEntity.fromMap(groupData);
-    } else {
-      throw FirebaseException(
-        plugin: 'Firebase',
-        message: 'Group does not exist',
-      );
-    }
-  }
-
-  @override
-  Future<SubGroupEntity> findSubGroup(String subGroupId) async {
-    final snapshot = await FirebaseFirestore.instance
-        .collection('subGroups')
-        .doc(subGroupId)
-        .get();
-    return SubGroupEntity.fromMap(snapshot.data()!);
-  }
-
-  @override
-  Future<List<GroupEntity>> findCreatedGroups(String userId) async {
-    final snapshot = await FirebaseFirestore.instance
-        .collection('groups')
-        .where('createdBy', isEqualTo: userId)
-        .get();
-    final results =
-        snapshot.docs.map((e) => GroupEntity.fromMap(e.data())).toList();
-    return results;
-  }
-
-  @override
-  Future<List<SubGroupEntity>> findJoinedGroups(
-      GroupMemberEntity member) async {
-    // find all the sub groups where user is a member
-
-    final snapshot = await FirebaseFirestore.instance
-        .collection('subGroups')
-        .where('members', arrayContains: member.toMap())
-        .get();
-    final results =
-        snapshot.docs.map((e) => SubGroupEntity.fromMap(e.data())).toList();
-    return results;
-  }
-}
