--- lib/app/features/groups/presentation/controllers/created_communities_controller.dart
@@ -1,34 +1,0 @@
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:get/get.dart';
-import 'package:organizer_client/app/features/groups/domain/entities/group_entity.dart';
-import 'package:organizer_client/app/features/groups/domain/usecases/find_created_groups.dart';
-import 'package:organizer_client/shared/ui/error_snackbar.dart';
-import 'package:organizer_client/shared/usecase/usecase.dart';
-
-class CreatedCommunitiesController extends GetxController {
-  RxBool isEmpty = true.obs;
-  RxBool isLoading = false.obs;
-  final FindCreatedGroupsUseCase findCreatedGroupsUseCase;
-  late final List<GroupEntity> groups;
-  CreatedCommunitiesController({required this.findCreatedGroupsUseCase});
-
-  @override
-  void onInit() {
-    findCreatedGroups();
-    super.onInit();
-  }
-
-  Future<void> findCreatedGroups() async {
-    isLoading.value = true;
-    final result = await findCreatedGroupsUseCase.call(
-      StringParams(FirebaseAuth.instance.currentUser!.uid),
-    );
-    result.fold((failure) {
-      showErrorSnackbar(message: failure.message);
-    }, (success) {
-      groups = success;
-      isEmpty.value = success.isEmpty;
-    });
-    isLoading.value = false;
-  }
-}
