+++ lib/app/routes/app_pages.dart
@@ -1,15 +1,15 @@
 import 'package:get/get.dart';
 import 'package:organizer_client/app/features/account/presentation/bindings/account_binding.dart';
+import 'package:organizer_client/app/features/community/presentation/bindings/create_communities_binding.dart';
+import 'package:organizer_client/app/features/community/presentation/bindings/groups_binding.dart';
+import 'package:organizer_client/app/features/community/presentation/bindings/new_group_binding.dart';
+import 'package:organizer_client/app/features/community/presentation/pages/created_communities_page.dart';
+import 'package:organizer_client/app/features/community/presentation/pages/new_group_page.dart';
 import 'package:organizer_client/app/features/discover/presentation/bindings/discover_binding.dart';
 import 'package:organizer_client/app/features/discover/presentation/bindings/join_group_bindng.dart';
 import 'package:organizer_client/app/features/discover/presentation/bindings/sub_group_binding.dart';
 import 'package:organizer_client/app/features/discover/presentation/pages/join_group_page.dart';
 import 'package:organizer_client/app/features/discover/presentation/pages/sub_group_page.dart';
-import 'package:organizer_client/app/features/groups/presentation/bindings/create_communities_binding.dart';
-import 'package:organizer_client/app/features/groups/presentation/bindings/groups_binding.dart';
-import 'package:organizer_client/app/features/groups/presentation/bindings/new_group_binding.dart';
-import 'package:organizer_client/app/features/groups/presentation/pages/created_communities_page.dart';
-import 'package:organizer_client/app/features/groups/presentation/pages/new_group_page.dart';
 import 'package:organizer_client/app/features/home/presentation/bindings/home_binding.dart';
 import 'package:organizer_client/app/features/home/presentation/pages/home_page.dart';
 import 'package:organizer_client/app/features/register/presentation/bindings/register_binding.dart';
