+++ lib/shared/constant/db_collections.dart
@@ -1,0 +1,6 @@
+// ignore_for_file: constant_identifier_names
+
+const COMMUNITIES_COLLECTION = "communities";
+const GROUPS_COLLECTION = 'groups';
+const USERS_COLLECTION = 'users';
+const MEMBERS_COLLECTION = "members";
