+++ lib/shared/enums/id.dart
@@ -1,0 +1,1 @@
+enum IdType { community, group }
